# docs/swagger/libraryapi_swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Webtor API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Who the key belongs to, which plan the account is on, what the key is allowed to do, and the\npreferences that affect what the other endpoints return.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Account state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partial update: a field left out of the body keeps its current value. Requires the `api:write`\nscope.\n\nPreferences are read on the hot path and cached for a few minutes, so a change can take that long\nto show up in poster URLs — this endpoint invalidates its own cache entry, but other replicas\nexpire on their own schedule.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update account preferences",
                "parameters": [
                    {
                        "description": "Fields to change",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "libapi.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "the specified path does not exist"
                }
            }
        },
        "libapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/libapi.Error"
                }
            }
        },
...
        "libapi.ProfileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "scopes": {
                    "description": "Scopes are what the key used for this request may do.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "api:read",
                        "api:write"
                    ]
                },
                "settings": {
                    "$ref": "#/definitions/libapi.ProfileSettings"
                },
                "tier": {
                    "$ref": "#/definitions/libapi.ProfileTier"
                },
                "user_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
        "libapi.ProfileSettings": {
            "type": "object",
            "properties": {
                "show_adult": {
                    "description": "ShowAdult opts out of the server-side blur on adult content.",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "libapi.ProfileSettingsRequest": {
            "type": "object",
            "properties": {
                "show_adult": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "libapi.ProfileTier": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Pro"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API key from your profile page, as `Bearer \u003ckey\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}